Remove cesium.usd.plugins as separate extension
