+++ .vscode/cesium-omniverse-linux.code-workspace
@@ -30,7 +30,6 @@
     // Python modules search paths:
     "python.analysis.extraPaths": [
       "${workspaceFolder}/exts/cesium.omniverse",
-      "${workspaceFolder}/exts/cesium.usd.plugins",
       "${workspaceFolder}/extern/nvidia/_build/target-deps/kit-sdk/kernel/py",
       "${workspaceFolder}/extern/nvidia/_build/target-deps/kit-sdk/exts/carb.audio",
       "${workspaceFolder}/extern/nvidia/_build/target-deps/kit-sdk/exts/carb.hsscclient",
