+++ apps/cesium.omniverse.dev.kit
@@ -25,6 +25,6 @@ app.fastShutdown = true
 folders.'++' = [
     "${app}", # Find other applications in this folder
     "${app}/exts", # Find extensions in this folder
-    "${app}/../exts", # Find cesium.omniverse and cesium.usd.schemas
+    "${app}/../exts", # Find cesium.omniverse
     "${app}/../extern/nvidia/app/extscache" # Find omni.kit.window.material_graph
 ]
