+++ apps/cesium.omniverse.dev.python.debug.kit
@@ -15,6 +15,6 @@ app.window.title = "Cesium For Omniverse Python Debugging App"
 folders.'++' = [
     "${app}", # Find other applications in this folder
     "${app}/exts", # Find extensions in this folder
-    "${app}/../exts", # Find cesium.omniverse and cesium.usd.schemas
+    "${app}/../exts", # Find cesium.omniverse
     "${app}/../extern/nvidia/app/extscache" # Find omni.kit.window.material_graph
 ]
