+++ exts/cesium.omniverse/config/extension.toml
@@ -33,7 +33,6 @@ python = false
 
 # Which extensions this extension depends on
 [dependencies]
-"cesium.usd.plugins" = { version = "0.3.0" }
 "usdrt.scenegraph" = {}
 "omni.ui" = {}
 "omni.usd" = {}
