--- exts/cesium.usd.plugins/config/extension.toml
@@ -1,49 +1,0 @@
-[core]
-# Load after omni.usd.libs (-1000) and before omni.usd (0)
-# See https://docs.omniverse.nvidia.com/py/kit/docs/guide/usd_schema.html
-order = -100
-
-[package]
-version = "0.3.0"
-category = "simulation"
-feature = false
-app = false
-title = "Cesium for Omniverse USD Plugins"
-description = "Supporting USD Plugins for Cesium for Omniverse"
-authors = "Cesium GS Inc."
-repository = "https://github.com/CesiumGS/cesium-omniverse"
-keywords = [
-    "cesium",
-    "omniverse",
-    "geospatial",
-    "3D Tiles",
-    "glTF",
-    "globe",
-    "earth",
-    "simulation",
-]
-toggleable = false
-
-# Paths are relative to the extension folder
-changelog = "doc/CHANGES.md"
-readme = "doc/README.md"
-preview_image = "doc/images/preview.jpg"
-icon = "doc/images/icon.png"
-
-[package.target]
-kit = ["105.1"]
-
-[package.writeTarget]
-kit = true
-python = false
-
-# Which extensions this extension depends on
-[dependencies]
-"omni.usd.libs" = {}
-
-# Main python module this extension provides, it will be publicly available as "import cesium.usd.plugins"
-[[python.module]]
-name = "cesium.usd.plugins"
-
-[[native.library]]
-path = "bin/${lib_prefix}CesiumUsdSchemas${lib_ext}"
