--- exts/cesium.usd.plugins/doc/CHANGES.md
@@ -1,5 +1,0 @@
-# Change Log
-
-### v0.1.0 - 2023-03-20
-
-- Split the Cesium USD plugins into a separate extension.
