--- exts/cesium.usd.plugins/doc/README.md
@@ -1,3 +1,0 @@
-# Cesium for Omniverse USD Plugins
-
-Supporting package for Cesium for Omniverse containing Cesium's USD plugins and schemas.
