+++ regenerate_schema.bat
@@ -9,7 +9,7 @@ set NVIDIA_USD_PYTHON_LIBS=%NVIDIA_USD_ROOT%\lib\python
 set NVIDIA_PYTHON_ROOT=%PROJECT_ROOT%\extern\nvidia\_build\target-deps\python
 set NVIDIA_PYTHON_EXECUTABLE=%NVIDIA_PYTHON_ROOT%\python.exe
 
-set SCHEMA_INPUT_PATH=%PROJECT_ROOT%\exts\cesium.usd.plugins\schemas\cesium_schemas.usda
+set SCHEMA_INPUT_PATH=%PROJECT_ROOT%\exts\cesium.omniverse\schemas\cesium_schemas.usda
 set SCHEMA_OUTPUT_PATH=%PROJECT_ROOT%\src\plugins\CesiumUsdSchemas\src\CesiumUsdSchemas
 
 set PYTHONPATH=%NVIDIA_USD_PYTHON_LIBS%;%PYTHONPATH%
